requirements.txt
# It's easy to add more libraries or choose different versions. Any libraries
# specified here will be installed and made available to your morph.io scraper.
# Find out more: https://morph.io/documentation/python

# Custom version of scraperwiki library
#-e git+http://github.com/openaustralia/scraperwiki-#python.git@morph_defaults#egg=scraperwiki

lxml==3.4.4
cssselect==0.9.1
beautifulsoup4==4.4.1
fuzzywuzzy==0.10.0
python-Levenshtein==0.12.0
requests==2.9.1
python-dateutil==2.5.3

